added boss spawn delay variable to spawn manager, changed sun color in levels 3/4, changed way kamikaze drone does damage

+++ Assets/Managers/Resource Manager.cs
@@ -26,9 +26,9 @@ public class ResourceManager : MonoBehaviour
     }
 
     //Increase the current amount of resources by the passed parameter
-    public void IncreaseResources(int value)
+    public void IncreaseResources(float value)
     {
-        currentResources += value;
+        currentResources += (int)value;
         resourceText.SetText(currentResources.ToString());
         resourceTextAnimator.SetTrigger("Pulse");
     }
